pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "model-belief"
version = "0.1.0"
description = "Extract LLM model beliefs from token-level probabilities"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
  "openai>=2.16.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "integration: tests that call external services (opt-in)"
]